refactor(data-model): move the JSON convenience layer to `codecs.ts`
Second of three preparing `codec-*` for `codec-realm`. A relocation: no
behavior changes and no signature does.

`jsonFromValue()`, `valueFromJson()` and `plainObjectFromJson()` are not the
JSON format. They are the JSON format bound to one set of fabric classes and
one reconstruction context, which is a layer above it. They move to a new
`data-model/src/codecs.ts`, along with the two module-level singletons they
share -- the codec instance and the decode-framed empty context -- so the
package's ambient codec defaults sit in one file.

`seemsLikeJsonEncodedFabricValue()` stays behind in `codec-json/impl.ts`,
because it only sniffs for the format's prefix tag and so belongs to the
format. Its doc comment said "as defined by this module", which named a module
that now defines nothing; it names `JsonCodec` instead.

`impl.test.ts` splits the same way, with its `seemsLike...` cases staying and
the rest becoming `test/codecs.test.ts`. Step count is unchanged across the
split (2261), which is the check that no case was dropped.

The formal spec described these functions as living in a dedicated module and
gave the path; it now describes the two layers and gives both paths. Its path
had also lost the `src/` component, fixed in passing.

Sixteen import sites re-point to the new `"./codecs"` subpath. One of them is
`tasks/pattern-compat-lib.ts`, outside `packages/` -- found by `deno task
check` after a first sweep scoped to `packages/` missed it.
