+++ packages/data-model/deno.jsonc
@@ -39,6 +39,7 @@
     "./value-clone": "./src/value-clone.ts",
     "./frozen-builtins": "./src/frozen-builtins.ts",
     "./codec-json": "./src/codec-json/index.ts",
+    "./codecs": "./src/codecs.ts",
     "./data-uri-codec": "./src/data-uri-codec.ts",
     "./native-type-tags": "./src/native-type-tags.ts",
     "./schema-hash": "./src/schema-hash.ts",
