+++ packages/runner/test/memory-v2-stacked-commit.test.ts
@@ -4,10 +4,7 @@ import {
   assertExists,
   assertStrictEquals,
 } from "@std/assert";
-import {
-  jsonFromValue,
-  valueFromJson,
-} from "@commonfabric/data-model/codec-json";
+import { jsonFromValue, valueFromJson } from "@commonfabric/data-model/codecs";
 import { FabricEpochNsec } from "@commonfabric/data-model/fabric-primitives";
 import { Identity } from "@commonfabric/identity";
 import type { FabricValue } from "@commonfabric/api";
